refactor api serializers
