+++ backend/core/serializers/relationserializer.py
@@ -5,7 +5,6 @@ from varname import nameof
 # Local libraries
 from .. models import *
 from ..serializers.authorserializer import AuthorSerializer
-from ..api_serializers import *
 
 author_serializer = AuthorSerializer()
