+++ backend/core/urls.py
@@ -4,7 +4,8 @@ from drf_yasg import openapi
 from django.urls import path, re_path
 from django.shortcuts import redirect, render
 
-from . import views
+from . import views1
+from .views import account_views, author_views, follower_views, post_views, comment_views
 
 schema_view = get_schema_view(
    openapi.Info(
@@ -25,27 +26,32 @@ urlpatterns = [
     re_path(r'^swagger/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     re_path(r'^redoc/$', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
     
-    re_path(r"^$", views.index, name="index"),
+    # Account views
+    re_path(r"^$", account_views.index, name="index"),
+    path('signin/', account_views.signin, name='signin'),
+    path('signup/', account_views.signup, name='signup'),
+    path('accounts/logout/', account_views.logout_view, name='logout'),
+
+    # Author views
+    path('get_author_id/', author_views.CurrentAuthorID.as_view(), name="get_author_id"),
+    path('authors/', author_views.AuthorListAPI.as_view(), name='author_list'),
+    path('authors/<uuid:author_id>/', author_views.AuthorAPI.as_view(), name='author_api'),
+
+    # Follower views
+    path('authors/<uuid:author_id>/followers/', follower_views.FollowerListAPI.as_view(), name='followers_list'),
+    path('authors/<uuid:author_id>/followers/<uuid:foreign_author_id>/', follower_views.FollowerAPI.as_view(), name="follower_api"),
+
+    # Post views  
+    path('authors/<uuid:author_id>/posts/<uuid:post_id>/', post_views.SinglePostAPI.as_view(), name="post_single"),
+    path('authors/<uuid:author_id>/posts/', post_views.PostAPI.as_view(), name="post"),
+
+    # Comment views
+    path('authors/<uuid:author_id>/posts/<uuid:post_id>/comments/', comment_views.CommentAPI.as_view(), name="comment"),
     
-    path('get_author_id/', views.CurrentAuthorID.as_view(), name="get_author_id"),
+    # Like views
+    path('authors/<uuid:author_id>/posts/<uuid:post_id>/likes/', views1.LikeAPI.as_view(), name="like"),
+    path('authors/<uuid:author_id>/posts/<uuid:post_id>/comments/<uuid:comment_id>/likes/', views1.CommentLikeAPI.as_view(), name="comment_like"),
+    path('authors/<uuid:author_id>/inbox/', views1.InboxLike.as_view(), name="inbox_like"),
 
-    path('signin/', views.signin, name='signin'),
-    path('signup/', views.signup, name='signup'),
-    path('accounts/logout/', views.logout_view, name='logout'),
-    
-    path('authors/', views.AuthorListAPI.as_view(), name='author_list'),
-    path('authors/<uuid:author_id>/', views.AuthorAPI.as_view(), name='author_api'),
-    path('authors/<uuid:author_id>/followers/', views.FollowerListAPI.as_view(), name='followers_list'),
-    path('authors/<uuid:author_id>/followers/<uuid:foreign_author_id>/', views.FollowerAPI.as_view(), name="follower_api"),
-
-    path('authors/<uuid:author_id>/posts/<uuid:post_id>/', views.SinglePostAPI.as_view(), name="post_single"),
-    path('authors/<uuid:author_id>/posts/', views.PostAPI.as_view(), name="post"),
-
-    path('authors/<uuid:author_id>/posts/<uuid:post_id>/comments/', views.CommentAPI.as_view(), name="comment"),
-    
-    path('authors/<uuid:author_id>/posts/<uuid:post_id>/likes/', views.LikeAPI.as_view(), name="like"),
-    path('authors/<uuid:author_id>/posts/<uuid:post_id>/comments/<uuid:comment_id>/likes/', views.CommentLikeAPI.as_view(), name="comment_like"),
-    path('authors/<uuid:author_id>/inbox/', views.InboxLike.as_view(), name="inbox_like"),
-
-    path('authors/<uuid:author_id>/liked/', views.InboxLike.as_view(), name="author_liked")     
+    path('authors/<uuid:author_id>/liked/', views1.InboxLike.as_view(), name="author_liked")     
 ]
